test(multi-tenancy): notificationRule API for multi-tenancy (#5718)

+++ openaev-api/src/main/java/io/openaev/service/NotificationRuleService.java
@@ -2,6 +2,7 @@ package io.openaev.service;
 
 import static io.openaev.utils.pagination.PaginationUtils.buildPaginationJPA;
 
+import io.openaev.context.TenantContext;
 import io.openaev.database.model.*;
 import io.openaev.database.model.TenantSettingKeys;
 import io.openaev.database.repository.NotificationRuleRepository;
@@ -38,7 +39,7 @@ public class NotificationRuleService {
   private final TenantSettingsService tenantSettingsService;
 
   public Optional<NotificationRule> findById(final String id) {
-    return notificationRuleRepository.findById(id);
+    return notificationRuleRepository.findByIdAndTenantId(id, TenantContext.getCurrentTenant());
   }
 
   public List<NotificationRule> findAll() {
@@ -79,7 +80,7 @@ public class NotificationRuleService {
     // verify that the rule exists
     NotificationRule notificationRule =
         notificationRuleRepository
-            .findById(id)
+            .findByIdAndTenantId(id, TenantContext.getCurrentTenant())
             .orElseThrow(
                 () -> new ElementNotFoundException("NotificationRule not found with id: " + id));
 
@@ -89,12 +90,9 @@ public class NotificationRuleService {
   }
 
   public void deleteNotificationRule(@NotBlank final String id) {
-    // verify that the rule exists
-    notificationRuleRepository
-        .findById(id)
-        .orElseThrow(
-            () -> new ElementNotFoundException("NotificationRule not found with id: " + id));
-
+    if (!notificationRuleRepository.existsByIdAndTenantId(id, TenantContext.getCurrentTenant())) {
+      throw new ElementNotFoundException("NotificationRule not found with id: " + id);
+    }
     notificationRuleRepository.deleteById(id);
   }
